python-back/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
# psycopg2-binary==2.9.9  # Not needed for SQLite
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<5.0.0  # passlib 1.7.4 requires bcrypt 4.x

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Scientific computing (for reliability calculations)
scipy>=1.13.0
sympy==1.12

# Utilities
python-dotenv==1.0.0
